+++ display/paint_driver/fonts/fonts.h
@@ -97,6 +97,8 @@ extern const Font font_20_cyrillic;
 extern const Font font_24_cyrillic;
 extern const Font font_32_cyrillic;
 extern const Font font_36_cyrillic;
+extern const Font font_40_cyrillic;
+extern const Font font_48_cyrillic;
 extern const Font font_8_ascii;
 extern const Font font_12_ascii;
 extern const Font font_16_ascii;
@@ -104,6 +106,8 @@ extern const Font font_20_ascii;
 extern const Font font_24_ascii;
 extern const Font font_32_ascii;
 extern const Font font_36_ascii;
+extern const Font font_40_ascii;
+extern const Font font_48_ascii;
 
 extern cFONT Font12CN;
 extern cFONT Font24CN;