+++ webserver/pages/root.cpp
@@ -2,7 +2,7 @@
 #include "../../intl.h"
 #include "../html-content.h"
 
-void webserver_root(String &page_content) {
+void webserver_root(String &page_content, const String &robonomics_address) {
 
     // Enable Pagination
     page_content += FPSTR(WEB_ROOT_PAGE_CONTENT);
@@ -11,4 +11,11 @@ void webserver_root(String &page_content) {
     page_content.replace(F("{conf}"), FPSTR(INTL_CONFIGURATION));
     page_content.replace(F("{restart}"), FPSTR(INTL_RESTART_SENSOR));
     page_content.replace(F("{debug}"), FPSTR(INTL_DEBUG_LEVEL));
+
+    String map_url = F("https://sensors.social/");
+    if (robonomics_address.length() > 0 && robonomics_address != F("Not Set")) {
+        map_url += F("?sensor=");
+        map_url += robonomics_address;
+    }
+    page_content.replace(F("https://sensors.social/"), map_url);
 }